refactor(functions): data-integrity endpointを薄いハンドラ+オーケストレータ構成へ再配置 (SPR-231 段階④・最終)
624行のdata-integrity-check.tsをendpoints/data-integrity/の3ファイルへ
分割し、SPR-231の全段階を完了する（挙動不変・移動のみ）:

- check-data-integrity.ts: 薄いハンドラ（CloudEvent受領→オーケストレータ
  呼び出し→結果ログ）
- run-integrity-check.ts: オーケストレータ（4検査の実行順序・結果集計・
  保存。tools/テストの単一import元として検査側シンボルを再エクスポート）
- integrity-checks.ts: 4種の検査・修復関数（中身は一字も変更なし）

整合性cron領域（One-Way Door）のためロジック・ログ文言・定数値は完全に
逐語移動。リネームなし・Cloud Functions登録名checkDataIntegrityは不変。
追従: index.ts import 1行・tools/core import 1行・package.jsonの
test:integrationパス・biome.jsonの認知的複雑度除外glob（検査関数の
移動先へ）・database-schema.mdリンク3箇所。テストは
endpoints/data-integrity/__tests__/へ移設（importパスのみ変更・
アサーション無変更）。
